Address PR 115 post-rebase review threads
Resolves review thread PRRT_kwDORS2O8s6DnNUp

Resolves review thread PRRT_kwDORS2O8s6DnNUs

Resolves review thread PRRT_kwDORS2O8s6DnNUv
